Fix compilation errors in EtherealMind mod
- Add missing getItem/setItem/getTotalPages methods to DimensionalStorage
- Add extractItem overload that accepts ItemStack parameter
- Fix getTarget() cast issue in EnvironmentAwareness (cast to Mob)
- Change CROP_HARVEST_RANGE from double to int
- Fix getCurrentMood() to use cosmo.getMood() instead

+++ src/main/java/com/astrolabs/etherealmind/common/entity/behavior/ResourceGatherer.java
@@ -39,7 +39,7 @@ public class ResourceGatherer {
     
     // Gathering ranges
     private static final double ITEM_COLLECT_RANGE = 8.0;
-    private static final double CROP_HARVEST_RANGE = 6.0;
+    private static final int CROP_HARVEST_RANGE = 6;
     private static final double XP_COLLECT_RANGE = 10.0;
     
     // Statistics
